+++ apps/flows/tsconfig.json
@@ -43,7 +43,8 @@
       "@kirbydesign/designsystem/slide": ["libs/designsystem/slide/index.ts"],
       "@kirbydesign/designsystem/toggle-button": ["libs/designsystem/toggle-button/index.ts"],
       "@kirbydesign/designsystem/tabs": ["libs/designsystem/tabs/index.ts"],
-      "@kirbydesign/designsystem/slide-button": ["libs/designsystem/slide-button/index.ts"]
+      "@kirbydesign/designsystem/slide-button": ["libs/designsystem/slide-button/index.ts"],
+      "@kirbydesign/designsystem/range": ["libs/designsystem/range/index.ts"]
     },
     "target": "es2020"
   },
