+++ libs/designsystem/range/src/range.component.scss
