+++ libs/designsystem/range/src/range.component.ts
