+++ libs/designsystem/src/lib/index.ts
@@ -17,6 +17,7 @@ export * from '@kirbydesign/designsystem/toggle';
 export * from '@kirbydesign/designsystem/toggle-button';
 export * from '@kirbydesign/designsystem/types';
 export * from '@kirbydesign/designsystem/slide-button';
+export * from '@kirbydesign/designsystem/range';
 
 export * from './scss/scss-helper';
