+++ libs/designsystem/src/lib/kirby.module.ts
@@ -19,6 +19,7 @@ import { SlideModule } from '@kirbydesign/designsystem/slide';
 import { ToggleButtonModule } from '@kirbydesign/designsystem/toggle-button';
 import { TabsModule } from '@kirbydesign/designsystem/tabs';
 import { SlideButtonComponent } from '@kirbydesign/designsystem/slide-button';
+import { RangeComponent } from '@kirbydesign/designsystem/range';
 import { AccordionItemComponent } from './components/accordion/accordion-item.component';
 import { AccordionDirective } from './components/accordion/accordion.directive';
 import { AppModule } from './components/app/app.module';
@@ -56,7 +57,6 @@ import { PageModule } from './components/page/page.module';
 import { PopoverComponent } from './components/popover/popover.component';
 import { ProgressCircleRingComponent } from './components/progress-circle/progress-circle-ring.component';
 import { ProgressCircleComponent } from './components/progress-circle/progress-circle.component';
-import { RangeComponent } from './components/range/range.component';
 import { ReorderListComponent } from './components/reorder-list/reorder-list.component';
 import { RouterOutletModule } from './components/router-outlet/router-outlet.module';
 import { SegmentedControlComponent } from './components/segmented-control/segmented-control.component';
