+++ libs/designsystem/tsconfig.json
@@ -19,7 +19,9 @@
       "@kirbydesign/designsystem/slide": ["slide/index.ts"],
       "@kirbydesign/designsystem/toggle-button": ["toggle-button/index.ts"],
       "@kirbydesign/designsystem/tabs": ["tabs/index.ts"],
-      "@kirbydesign/designsystem/slide-button": ["slide-button/index.ts"]
+      "@kirbydesign/designsystem/slide-button": ["slide-button/index.ts"],
+      "@kirbydesign/designsystem/range": ["range/index.ts"]
+
     },
     "esModuleInterop": true,
     "target": "es2020"
