+++ libs/designsystem/tsconfig.spec.json
@@ -22,7 +22,8 @@
       "@kirbydesign/designsystem/slide": ["slide/index.ts"],
       "@kirbydesign/designsystem/toggle-button": ["toggle-button/index.ts"],
       "@kirbydesign/designsystem/tabs": ["tabs/index.ts"],
-      "@kirbydesign/designsystem/slide-button": ["slide-button/index.ts"]
+      "@kirbydesign/designsystem/slide-button": ["slide-button/index.ts"],
+      "@kirbydesign/designsystem/range": ["range/index.ts"]
     }
   },
   "files": ["test.ts"],
